move all tablature styles to Tablature.module.scss
