--- src/modules/tablatureEditor/components/Column.module.scss
@@ -1,16 +1,0 @@
-.column {
-	line-height: 10px;
-	writing-mode: vertical-lr;
-	text-orientation: upright;
-
-	&[data-selected='true'],
-	&[data-ghost-selected='true'] {
-		color: $c-background;
-		background-color: $c-foreground;
-	}
-
-	&:hover {
-		background-color: $c-medium;
-		cursor: pointer;
-	}
-}
