--- src/modules/tablatureEditor/components/Line.module.scss
@@ -1,7 +1,0 @@
-.section {
-	display: flex;
-	flex-wrap: wrap;
-	margin: 0.5rem 0;
-	padding: 0.4rem 1rem 0.4rem 0.5rem;
-	box-shadow: 1px 2px 7px 0 #9b9b9b;
-}
