+++ src/modules/tablatureEditor/components/Tablature.module.scss
@@ -5,3 +5,28 @@
 	font-size: 12px;
 	user-select: none;
 }
+
+.section {
+	display: flex;
+	flex-wrap: wrap;
+	margin: 0.5rem 0;
+	padding: 0.4rem 1rem 0.4rem 0.5rem;
+	box-shadow: 1px 2px 7px 0 #9b9b9b;
+}
+
+.column {
+	line-height: 10px;
+	writing-mode: vertical-lr;
+	text-orientation: upright;
+
+	&[data-selected='true'],
+	&[data-ghost-selected='true'] {
+		color: $c-background;
+		background-color: $c-foreground;
+	}
+
+	&:hover {
+		background-color: $c-medium;
+		cursor: pointer;
+	}
+}
